refactor: switch to float-based 2D math (#3)

+++ libs/ui/include/seir_ui/margins.hpp
@@ -9,19 +9,19 @@ namespace seir
 	class Margins
 	{
 	public:
-		int _top = 0;
-		int _right = 0;
-		int _bottom = 0;
-		int _left = 0;
+		float _top = 0;
+		float _right = 0;
+		float _bottom = 0;
+		float _left = 0;
 
 		constexpr Margins() noexcept = default;
-		constexpr explicit Margins(int all) noexcept
+		constexpr explicit Margins(float all) noexcept
 			: _top{ all }, _right{ all }, _bottom{ all }, _left{ all } {}
-		constexpr Margins(int topBottom, int leftRight) noexcept
+		constexpr Margins(float topBottom, float leftRight) noexcept
 			: _top{ topBottom }, _right{ leftRight }, _bottom{ topBottom }, _left{ leftRight } {}
-		constexpr Margins(int top, int leftRight, int bottom) noexcept
+		constexpr Margins(float top, float leftRight, float bottom) noexcept
 			: _top{ top }, _right{ leftRight }, _bottom{ bottom }, _left{ leftRight } {}
-		constexpr Margins(int top, int right, int bottom, int left) noexcept
+		constexpr Margins(float top, float right, float bottom, float left) noexcept
 			: _top{ top }, _right{ right }, _bottom{ bottom }, _left{ left } {}
 	};
 }
